+++ vs2019/simpleprint_vs2010_vs2019/simpleprint_vs2010_vs2019.cpp
@@ -8,7 +8,7 @@ int _tmain(int argc, TCHAR* argv[])
 	setlocale(LC_ALL, "");
 	
 	_tprintf(_T("Hello, simpleprint_vs2010_vs2019!\n"));
-	_tprintf(_T("sizeof(TCHAR)=%d\n"), sizeof(TCHAR));
+	_tprintf(_T("sizeof(TCHAR)=%d\n"), (int)sizeof(TCHAR));
 	return 0;
 }
